Move shared functionality to new API class
This commit is designed to have no significant changes except moving the API calls that are used by the newly shared components into a new common PublishApi.
