+++ src/BloomExe/Publish/Android/PublishToAndroidApi.cs
@@ -17,6 +17,7 @@ using Bloom.Publish.Android.usb;
 #endif
 using Bloom.Publish.Android.wifi;
 using Bloom.web;
+using Bloom.web.controllers;
 using BloomTemp;
 using DesktopAnalytics;
 using SIL.IO;
@@ -41,31 +42,18 @@ namespace Bloom.Publish.Android
 		private readonly BookServer _bookServer;
 		private readonly BulkBloomPubCreator _bulkBloomPubCreator;
 		private readonly WebSocketProgress _progress;
-		private const string kWebSocketContext = "publish-android"; // must match what is in AndroidPublishUI.tsx
-		private Color _thumbnailBackgroundColor = Color.Transparent; // can't be actual book cover color <--- why not?
-		private Book.Book _coverColorSourceBook;
 
-		private object _lockForLanguages = new object();
-		private Dictionary<string, bool> _allLanguages;
-		private HashSet<string> _languagesWithAudio = new HashSet<string>();
-		private Bloom.Book.Book _bookForLanguagesToPublish = null;
-
-		// This constant must match the ID that is used for the listener set up in the React component AndroidPublishUI
-		private const string kWebsocketEventId_Preview = "androidPreview";
-		public const string StagingFolder = "PlaceForStagingBook";
-
-		// This constant must match the ID used for the useWatchString called by the React component MethodChooser.
-		private const string kWebsocketState_LicenseOK = "publish/licenseOK";
-
-		public static string PreviewUrl { get; set; }
+		private PublishApi _publishApi;
 		
-		public PublishToAndroidApi(CollectionSettings collectionSettings, BloomWebSocketServer bloomWebSocketServer, BookServer bookServer, BulkBloomPubCreator bulkBloomPubCreator)
+		public PublishToAndroidApi(CollectionSettings collectionSettings, BloomWebSocketServer bloomWebSocketServer, BookServer bookServer, BulkBloomPubCreator bulkBloomPubCreator,
+			PublishApi publishApi)
 		{
 			_collectionSettings = collectionSettings;
 			_webSocketServer = bloomWebSocketServer;
 			_bookServer = bookServer;
 			_bulkBloomPubCreator = bulkBloomPubCreator;
-			_progress = new WebSocketProgress(_webSocketServer, kWebSocketContext);
+			_publishApi = publishApi;
+			_progress = new WebSocketProgress(_webSocketServer, PublishApi.kWebSocketContext);
 			_wifiPublisher = new WiFiPublisher(_progress, _bookServer);
 #if !__MonoCS__
 			_usbPublisher = new UsbPublisher(_progress, _bookServer)
@@ -75,25 +63,6 @@ namespace Bloom.Publish.Android
 #endif
 		}
 
-		/// <summary>
-		/// Conceptually, this is where we are currently building a book for preview.
-		/// In the current implementation, it is not cleared when we are no longer doing so.
-		/// Nor does it include the path to the individual book folder, just the staging
-		/// folder. This is not ideal but it serves the current limited purpose of this field.
-		/// </summary>
-		public static string CurrentPublicationFolder { get; private set; }
-		
-
-		private static string ToCssColorString(System.Drawing.Color c)
-		{
-			return "#" + c.R.ToString("X2") + c.G.ToString("X2") + c.B.ToString("X2");
-		}
-
-		private AndroidPublishSettings GetSettings()
-		{
-			return AndroidPublishSettings.FromBookInfo(_bookForLanguagesToPublish.BookInfo);
-		}
-		
 		public void RegisterWithApiHandler(BloomApiHandler apiHandler)
 		{
 			// This is just for storing the user preference of method
@@ -123,80 +92,18 @@ namespace Bloom.Publish.Android
 				}
 			}, true);
 
-			apiHandler.RegisterEndpointHandler(kApiUrlPart + "backColor", request =>
-			{
-				if (request.HttpMethod == HttpMethods.Get)
-				{
-					if (request.CurrentBook != _coverColorSourceBook)
-					{
-						_coverColorSourceBook = request.CurrentBook;
-						ImageUtils.TryCssColorFromString(request.CurrentBook?.GetCoverColor()??"", out _thumbnailBackgroundColor);
-					}
-					request.ReplyWithText(ToCssColorString(_thumbnailBackgroundColor));
-				}
-				else // post
-				{
-					// ignore invalid colors (very common while user is editing hex)
-					Color newColor;
-					var newColorAsString = request.RequiredPostString();
-					if (ImageUtils.TryCssColorFromString(newColorAsString, out newColor))
-					{
-						_thumbnailBackgroundColor = newColor;
-						request.CurrentBook.SetCoverColor(newColorAsString);
-					}
-					request.PostSucceeded();
-				}
-			}, true);
-
-			apiHandler.RegisterBooleanEndpointHandler(kApiUrlPart + "motionBookMode",
-				readRequest =>
-				{
-					// If the user has taken off all possible motion, force not having motion in the
-					// Bloom Reader book.  See https://issues.bloomlibrary.org/youtrack/issue/BL-7680.
-					if (!readRequest.CurrentBook.HasMotionPages)
-						readRequest.CurrentBook.BookInfo.PublishSettings.BloomPub.Motion = false;
-					return readRequest.CurrentBook.BookInfo.PublishSettings.BloomPub.Motion;
-				},
-				(writeRequest, value) =>
-				{
-					writeRequest.CurrentBook.BookInfo.PublishSettings.BloomPub.Motion = value;
-					writeRequest.CurrentBook.BookInfo.SavePublishSettings();
-					_webSocketServer.SendEvent("publish", "motionChanged");
-				}
-			, true);
-
 			apiHandler.RegisterEndpointHandler(kApiUrlPart + "updatePreview", request =>
 			{
-				MakeBloompubPreview(request, false);
+				_publishApi.MakeBloompubPreview(request, false);
 			}, false);
-
-			apiHandler.RegisterEndpointLegacy(kApiUrlPart + "thumbnail", request =>
-			{
-				var coverImage = request.CurrentBook.GetCoverImagePath();
-				if (coverImage == null)
-					request.Failed("no cover image");
-				else
-				{
-					// We don't care as much about making it resized as making its background transparent.
-					using(var thumbnail = TempFile.CreateAndGetPathButDontMakeTheFile())
-					{
-						if(_thumbnailBackgroundColor == Color.Transparent)
-						{
-							ImageUtils.TryCssColorFromString(request.CurrentBook?.GetCoverColor(), out _thumbnailBackgroundColor);
-						}
-						RuntimeImageProcessor.GenerateEBookThumbnail(coverImage, thumbnail.Path, 256, 256, _thumbnailBackgroundColor);
-						request.ReplyWithImage(thumbnail.Path);
-					}
-				}
-			}, true);
 
 			apiHandler.RegisterEndpointHandler(kApiUrlPart + "usb/start", request =>
 			{
 #if !__MonoCS__
 
 				SetState("UsbStarted");
-				UpdatePreviewIfNeeded(request);
-				_usbPublisher.Connect(request.CurrentBook, _thumbnailBackgroundColor, GetSettings());
+				_publishApi.UpdatePreviewIfNeeded(request);
+				_usbPublisher.Connect(request.CurrentBook, _publishApi._thumbnailBackgroundColor, _publishApi.GetSettings());
 #endif
 				request.PostSucceeded();
 			}, true);
@@ -209,31 +116,31 @@ namespace Bloom.Publish.Android
 #endif
 				request.PostSucceeded();
 			}, true);
-			apiHandler.RegisterEndpointLegacy(kApiUrlPart + "wifi/start", request =>
+			apiHandler.RegisterEndpointHandler(kApiUrlPart + "wifi/start", request =>
 			{
 				SetState("ServingOnWifi");
-				UpdatePreviewIfNeeded(request);
-				_wifiPublisher.Start(request.CurrentBook, request.CurrentCollectionSettings, _thumbnailBackgroundColor, GetSettings());
+				_publishApi.UpdatePreviewIfNeeded(request);
+				_wifiPublisher.Start(request.CurrentBook, request.CurrentCollectionSettings, _publishApi._thumbnailBackgroundColor, _publishApi.GetSettings());
 				
 				request.PostSucceeded();
 			}, true);
 
-			apiHandler.RegisterEndpointLegacy(kApiUrlPart + "wifi/stop", request =>
+			apiHandler.RegisterEndpointHandler(kApiUrlPart + "wifi/stop", request =>
 			{
 				_wifiPublisher.Stop();
 				SetState("stopped");
 				request.PostSucceeded();
 			}, true);
 
-			apiHandler.RegisterEndpointLegacy(kApiUrlPart + "file/save", request =>
+			apiHandler.RegisterEndpointHandler(kApiUrlPart + "file/save", request =>
 			{
-				UpdatePreviewIfNeeded(request);
-				FilePublisher.Save(request.CurrentBook, _bookServer, _thumbnailBackgroundColor, _progress, GetSettings());
+				_publishApi.UpdatePreviewIfNeeded(request);
+				FilePublisher.Save(request.CurrentBook, _bookServer, _publishApi._thumbnailBackgroundColor, _progress, _publishApi.GetSettings());
 				SetState("stopped");
 				request.PostSucceeded();
 			}, true);
 
-			apiHandler.RegisterEndpointLegacy(kApiUrlPart + "file/bulkSaveBloomPubsParams", request =>
+			apiHandler.RegisterEndpointHandler(kApiUrlPart + "file/bulkSaveBloomPubsParams", request =>
 			{ 
 				request.ReplyWithJson(JsonConvert.SerializeObject(_collectionSettings.BulkPublishBloomPubSettings));
 			}, true);
@@ -249,20 +156,11 @@ namespace Bloom.Publish.Android
 				request.PostSucceeded();
 			}, true);
 
-			apiHandler.RegisterEndpointLegacy(kApiUrlPart + "textToClipboard", request =>
+			apiHandler.RegisterEndpointHandler(kApiUrlPart + "textToClipboard", request =>
 			{
 				PortableClipboard.SetText(request.RequiredPostString());
 				request.PostSucceeded();
 			}, true);
-
-			apiHandler.RegisterBooleanEndpointHandler(kApiUrlPart + "canHaveMotionMode",
-				request =>
-				{
-					return request.CurrentBook.HasMotionPages;
-				},
-				null, // no write action
-				false,
-				true); // we don't really know, just safe default
 
 			apiHandler.RegisterBooleanEndpointHandler(kApiUrlPart + "canRotate",
 				request =>
@@ -281,232 +179,8 @@ namespace Bloom.Publish.Android
 				null, // no write action
 				false,
 				true); // we don't really know, just safe default
-			apiHandler.RegisterEndpointLegacy(kApiUrlPart + "languagesInBook", request =>
-			{
-				try
-				{
-					InitializeLanguagesInBook(request);
 
-					Dictionary<string, InclusionSetting> textLangsToPublish = request.CurrentBook.BookInfo.PublishSettings.BloomPub.TextLangs;
-					Dictionary<string, InclusionSetting> audioLangsToPublish = request.CurrentBook.BookInfo.PublishSettings.BloomPub.AudioLangs;
-					
-					var result = "[" + string.Join(",", _allLanguages.Select(kvp =>
-					{
-						string langCode = kvp.Key;
 
-						bool includeText = false;
-						if (textLangsToPublish != null && textLangsToPublish.TryGetValue(langCode, out InclusionSetting includeTextSetting))
-						{
-							includeText = includeTextSetting.IsIncluded();
-						}
-
-						bool includeAudio = false;
-						if (audioLangsToPublish != null && audioLangsToPublish.TryGetValue(langCode, out InclusionSetting includeAudioSetting))
-						{
-							includeAudio = includeAudioSetting.IsIncluded();
-						}
-
-						var value = new LanguagePublishInfo()
-						{
-							code = kvp.Key,
-							name = request.CurrentBook.PrettyPrintLanguage(langCode),
-							complete = kvp.Value,
-							includeText = includeText,
-							containsAnyAudio = _languagesWithAudio.Contains(langCode),							
-							includeAudio = includeAudio
-						};
-						var json = JsonConvert.SerializeObject(value);
-						return json;
-					})) + "]";
-
-					request.ReplyWithText(result);
-				}
-				catch (Exception e)
-				{
-					request.Failed("Error while determining languages in book. Message: " + e.Message);
-					NonFatalProblem.Report(ModalIf.Alpha, PassiveIf.All, "Error determining which languages are in the book.", null, e, true);
-				}
-			}, false);
-			apiHandler.RegisterEndpointLegacy(kApiUrlPart + "includeLanguage", request =>
-			{
-				var langCode = request.RequiredParam("langCode");
-				if (request.HttpMethod == HttpMethods.Post)
-				{
-					var includeTextValue = request.GetParamOrNull("includeText");
-					if (includeTextValue != null)
-					{
-						var inclusionSetting = includeTextValue == "true" ? InclusionSetting.Include : InclusionSetting.Exclude;
-						request.CurrentBook.BookInfo.PublishSettings.BloomPub.TextLangs[langCode] = inclusionSetting;
-					}
-
-					var includeAudioValue = request.GetParamOrNull("includeAudio");
-					if (includeAudioValue != null)
-					{
-						var inclusionSetting = includeAudioValue == "true" ? InclusionSetting.Include : InclusionSetting.Exclude;
-						request.CurrentBook.BookInfo.PublishSettings.BloomPub.AudioLangs[langCode] = inclusionSetting;
-					}
-
-					request.CurrentBook.BookInfo.Save();	// We updated the BookInfo, so need to persist the changes. (but only the bookInfo is necessary, not the whole book)
-					request.PostSucceeded();
-				}
-				// We don't currently need a get...it's subsumed in the 'include' value returned from allLanguages...
-				// but if we ever do this is what it would look like.
-				//else
-				//{
-				//	request.ReplyWithText(_languagesToPublish.Contains(langCode) ? "true" : "false");
-				//}
-			}, false);
-		}
-
-		public void MakeBloompubPreview(ApiRequest request, bool forVideo)
-		{
-			if (request.HttpMethod == HttpMethods.Post)
-			{
-				// This is already running on a server thread, so there doesn't seem to be any need to kick off
-				// another background one and return before the preview is ready. But in case something in C#
-				// might one day kick of a new preview, or we find we do need a background thread,
-				// I've made it a websocket broadcast when it is ready.
-				// If we've already left the publish tab...we can get a few of these requests queued up when
-				// a tester rapidly toggles between views...abandon the attempt
-				if (!PublishHelper.InPublishTab)
-				{
-					request.Failed("aborted, no longer in publish tab");
-					return;
-				}
-
-				try
-				{
-					UpdatePreview(request, forVideo);
-					request.PostSucceeded();
-				}
-				catch (Exception e)
-				{
-					request.Failed("Error while updating preview. Message: " + e.Message);
-					NonFatalProblem.Report(ModalIf.Alpha, PassiveIf.All, "Error while updating preview.", null, e, true);
-				}
-			}
-		}
-
-		private AndroidPublishSettings _lastSettings;
-		private Color _lastThumbnailBackgroundColor;
-
-		/// <summary>
-		/// The book language data needs to be initialized before handling updatePreview requests, but the
-		/// languagesInBook request comes in after the updatePreview request.  So we call this method in
-		/// both places with a lock to prevent stepping on each other.  This results in duplicate
-		/// calls for AllLanguages, but is safest since the user could leave the publish tab, change the
-		/// languages in the book, and then come back to the publish tab with the same book.
-		/// </summary>
-		private void InitializeLanguagesInBook(ApiRequest request)
-		{
-			lock (_lockForLanguages)
-			{
-				_allLanguages = request.CurrentBook.AllPublishableLanguages(includeLangsOccurringOnlyInXmatter: true);
-
-				// Note that at one point, we had a check that would bypass most of this function if the book hadn't changed.
-				// However, one side effect of this is that any settings behind the if guard would not be updated if the book was edited
-				// At one point, whenever a check box changed, the whole Publish screen was regenerated (along with languagesInBook being retrieved again),
-				// but this is no longer the case.
-				// So, we no longer have any bypass... Instead we recompute the values so that they can be updated
-				_bookForLanguagesToPublish = request.CurrentBook;
-
-				_languagesWithAudio = request.CurrentBook.GetLanguagesWithAudio();
-
-				InitializeLanguagesInBook(_bookForLanguagesToPublish.BookInfo, _allLanguages, request.CurrentCollectionSettings);
-			}
-		}
-
-		// Precondition: If any locking is required, the caller should handle it.
-		internal static void InitializeLanguagesInBook(BookInfo bookInfo, Dictionary<string, bool> allLanguages, CollectionSettings collectionSettings)
-		{
-			Debug.Assert(bookInfo?.MetaData != null, "Precondition: MetaData must not be null");
-
-			// reinitialize our list of which languages to publish, defaulting to the ones
-			// that are complete.
-			foreach (var kvp in allLanguages)
-			{
-				var langCode = kvp.Key;
-
-				// First, check if the user has already explicitly set the value. If so, we'll just use that value and be done.
-				if (bookInfo.PublishSettings.BloomPub.TextLangs.TryGetValue(langCode, out InclusionSetting checkboxValFromSettings))
-				{
-					if (checkboxValFromSettings.IsSpecified())
-					{
-						continue;
-					}
-				}
-
-				// Nope, either no value exists or the value was some kind of default value.
-				// Compute (or recompute) what the value should default to.
-				bool isChecked = kvp.Value ||
-					// We always select L1 by default because we assume the user wants to publish the language he is currently working on.
-					// It may be incomplete if he just wants to preview his work so far.
-					// If he really doesn't want to publish L1, he can deselect it.
-					// See BL-9587.
-					langCode == collectionSettings?.Language1Tag;
-
-				var newInitialValue = isChecked ? InclusionSetting.IncludeByDefault : InclusionSetting.ExcludeByDefault;
-				bookInfo.PublishSettings.BloomPub.TextLangs[langCode] = newInitialValue;
-			}
-
-			// Initialize the Talking Book Languages settings
-			if (bookInfo.PublishSettings.BloomPub.AudioLangs.Count == 0)
-			{
-				var allLangCodes = allLanguages.Select(x => x.Key);
-				foreach (var langCode in allLangCodes)
-				{
-					bookInfo.PublishSettings.BloomPub.AudioLangs[langCode] = InclusionSetting.IncludeByDefault;
-				}					
-			}
-
-			// The metadata may have been changed, so saved it.
-			// Note - If you want, you could check whether or not it was actually changed, but that might be premature optimization.
-			bookInfo.Save();
-		}
-
-		/// <summary>
-		/// Updates the BloomReader preview. The URL of the BloomReader preview will be sent over the web socket.
-		/// The format of the URL is a valid ("single" encoded) URL.
-		/// If the caller wants to insert this URL as a query parameter to another URL (e.g. like what is often done with Bloom Player),
-		/// it's the caller's responsibility to apply another layer of URL encoding to make the URL suitable to be passed as data inside another URL.
-		/// </summary>
-		private void UpdatePreview(ApiRequest request, bool forVideo)
-		{
-			InitializeLanguagesInBook(request);
-			_lastSettings = GetSettings();
-			_lastThumbnailBackgroundColor = _thumbnailBackgroundColor;
-			if (forVideo)
-			{
-				// We'll put all possible languages in the preview; the user can choose the one wanted
-				// with the preview controls if there is more than one. Don't include any that are NOT
-				// allowed to be published, or we won't get a preview at all, just an error message.
-				// Review: do we need to give some message about why some were not included?
-				var licenseChecker = new LicenseChecker();
-				var allowedLanguages = licenseChecker.AllowedLanguages(_allLanguages.Keys.ToArray(), request.CurrentBook);
-				if (!allowedLanguages.Any())
-				{
-					allowedLanguages = _allLanguages.Keys; // making preview will fail, but with a helpful message.
-				}
-				_lastSettings = _lastSettings.WithAllLanguages(allowedLanguages);
-			}
-
-			_lastSettings.Motion = forVideo ? request.CurrentBook.BookInfo.PublishSettings.AudioVideo.Motion : request.CurrentBook.BookInfo.PublishSettings.BloomPub.Motion;
-			_lastSettings.WantPageLabels = forVideo;
-			// BloomPlayer is capable of skipping these, but they confuse the page list we use to populate
-			// the page-range control.
-			_lastSettings.RemoveInteractivePages = forVideo;
-			PreviewUrl = MakeBloomPubForPreview(request.CurrentBook, _bookServer, _progress, _thumbnailBackgroundColor, _lastSettings);
-			_webSocketServer.SendString(kWebSocketContext, kWebsocketEventId_Preview, PreviewUrl);
-		}
-
-		private void UpdatePreviewIfNeeded(ApiRequest request)
-		{
-			var newSettings = GetSettings();
-			if (newSettings.Equals(_lastSettings) && _thumbnailBackgroundColor == _lastThumbnailBackgroundColor)
-			{
-				return;
-			}
-			UpdatePreview(request, false);
 		}
 
 		public void Dispose()
@@ -515,12 +189,11 @@ namespace Bloom.Publish.Android
 			_usbPublisher.Stop(disposing: true);
 #endif
 			_wifiPublisher.Stop();
-			_stagingFolder?.Dispose();
 		}
 
 		private void SetState(string state)
 		{
-			_webSocketServer.SendString(kWebSocketContext, kWebsocketState_EventId, state);
+			_webSocketServer.SendString(PublishApi.kWebSocketContext, kWebsocketState_EventId, state);
 		}
 
 		public static void ReportAnalytics(string mode, Book.Book book)
@@ -591,72 +264,6 @@ namespace Bloom.Publish.Android
 				progress.Message("PublishTab.Epub.Done", "Done", useL10nIdPrefix: false);	// share message string with epub publishing
 			}
 
-		}
-
-		private static TemporaryFolder _stagingFolder;
-
-		internal bool LicenseOK;
-
-		/// <summary>
-		/// Generates an unzipped, staged BloomPUB from the book
-		/// </summary>
-		/// <returns>A valid, well-formed URL on localhost that points to the staged book's htm file</returns>
-		public string MakeBloomPubForPreview(Book.Book book, BookServer bookServer, WebSocketProgress progress, Color backColor, AndroidPublishSettings settings = null)
-		{
-			progress.Message("PublishTab.Epub.PreparingPreview", "Preparing Preview");	// message shared with Epub publishing
-			if (settings?.LanguagesToInclude != null)
-			{
-				var message = new LicenseChecker().CheckBook(book, settings.LanguagesToInclude.ToArray());
-				if (message != null)
-				{
-					progress.MessageWithoutLocalizing(message, ProgressKind.Error);
-					LicenseOK = false;
-					_webSocketServer.SendString(kWebSocketContext, kWebsocketState_LicenseOK, "false");
-					return null;
-				}
-			}
-			LicenseOK = true;
-			_webSocketServer.SendString(kWebSocketContext, kWebsocketState_LicenseOK, "true");
-
-			_stagingFolder?.Dispose();
-			if (AudioProcessor.IsAnyCompressedAudioMissing(book.FolderPath, book.RawDom))
-			{
-				progress.Message("CompressingAudio", "Compressing audio files");
-				AudioProcessor.TryCompressingAudioAsNeeded(book.FolderPath, book.RawDom);
-			}
-			// BringBookUpToDate() will already have been done on the original book on entering the Publish tab.
-
-			// We don't use the folder found here, but this method does some checks we want done.
-			BookStorage.FindBookHtmlInFolder(book.FolderPath);
-			_stagingFolder = new TemporaryFolder(StagingFolder);
-			// I'd prefer this to include the book folder, but we need it before PrepareBookForBloomReader returns.
-			// I believe we only ever have one book being made there, so it works.
-			CurrentPublicationFolder = _stagingFolder.FolderPath;
-			var modifiedBook = BloomPubMaker.PrepareBookForBloomReader(settings, bookFolderPath: book.FolderPath, bookServer: bookServer, temp: _stagingFolder, progress,   isTemplateBook: book.IsTemplateBook);
-			progress.Message("Common.Done", "Shown in a list of messages when Bloom has completed a task.", "Done");
-			if (settings?.WantPageLabels ?? false)
-			{
-				int pageNum = 0;
-				var labelData
-					= modifiedBook.GetPages().Select(p =>
-					{
-						var caption = p.GetCaptionOrPageNumber(ref pageNum, out string i18nId);
-						if (!string.IsNullOrEmpty(caption))
-							caption = I18NApi.GetTranslationDefaultMayNotBeEnglish(i18nId, caption);
-						return caption;
-					}).ToArray();
-				dynamic messageBundle = new DynamicJson();
-				messageBundle.labels = labelData;
-				// We send these through the websocket rather than getting them through an API because the
-				// API request would very likely come before we finish generating the preview so at best we'd
-				// have to delay the response. Then the request might time out on a long book. So we'd really
-				// need an event anyway to tell the Typescript code it is time to request the labels.
-				// And then it would take some nasty spaghetti code to get the labels to the PublishVideoApi
-				// class so it could reply to the request. Cleaner just to send it through the socket.
-				_webSocketServer.SendBundle("publishPageLabels", "ready", messageBundle);
-			}
-
-			return modifiedBook.GetPathHtmlFile().ToLocalhost();
 		}
 
 		/// <summary>
